Made xPos floats, created a floor

+++ TrumpGame/src/com/tyler/gameObjects/GameObject.java
@@ -3,14 +3,14 @@ package com.tyler.gameObjects;
 public abstract class GameObject {
 
     // Variables
-    protected int x, y;
+    protected float x, y;
     protected int width, height;
     protected double velX, velY;
     protected ID id;
 
     // Constructor
     // TODO: Add Sprite Sheet
-    public GameObject (int x, int y, int width, int height, ID id) {
+    public GameObject (float x, float y, int width, int height, ID id) {
         this.x = x;
         this.y = y;
         this.width = width;
@@ -24,14 +24,14 @@ public abstract class GameObject {
     public abstract void physicsTick ();
 
     // Getters and Setters
-    public int getX () {
+    public float getX () {
         return x;
     }
     public void setX (int x) {
         this.x = x;
     }
 
-    public int getY () {
+    public float getY () {
         return y;
     }
     public void setY (int y) {
